refactor(lib): restore class files
