+++ docs/engineering.md
@@ -2,7 +2,7 @@
 
 ## Lexical block fields
 
-The `createOrUpdateHtmlFile` method of the [`payloadCrowdinSyncDocumentFilesApi`](plugin/src/lib/api/payload-crowdin-sync/files/document.ts) reads the collection config to determine whether the underlying field uses the [Slate editor](https://payloadcms.com/docs/rich-text/slate) or the [Lexical editor](https://payloadcms.com/docs/rich-text/lexical).
+The `createOrUpdateHtmlFile` method of the [`payloadCrowdinSyncDocumentFilesApi`](plugin/src/lib/api/files/document.ts) reads the collection config to determine whether the underlying field uses the [Slate editor](https://payloadcms.com/docs/rich-text/slate) or the [Lexical editor](https://payloadcms.com/docs/rich-text/lexical).
 
 If Lexical, blocks embedded within the editor are supported.
